Разбивает модуль utils на дополнительные модули
