+++ src/utils/common.js
@@ -34,30 +34,3 @@ export const uniteSeveralThingsIntoArray = (thing, amountThing) => {
 };
 
 export const getSomeValue = (arr) => arr[getRandomInteger(0, arr.length - 1)];
-
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-export const renderTemplate = (elem, where, html) => {
-  elem.insertAdjacentHTML(where, html);
-};
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
